requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
argcomplete==3.6.2
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
boto3==1.37.29
botocore==1.37.29
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
eval_type_backport==0.2.2
executing==2.2.0
fastapi==0.115.12
fastavro==1.10.0
filelock==3.18.0
flupy==1.2.1
frozenlist==1.5.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.69.2
gotrue==2.12.0
greenlet==3.1.1
griffe==1.7.2
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
logfire==3.12.0
logfire-api==3.12.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.6.0
mdurl==0.1.2
mem0ai==1.0.5
monotonic==1.6
multidict==6.3.2
narwhals==1.34.0
numpy==2.2.4
openai==1.71.0
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
packaging==24.2
pandas==2.2.3
pgvector==0.3.6
pillow==11.1.0
portalocker==2.10.1
postgrest==1.0.1
posthog==3.23.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-ai==1.39.0
pydantic-ai-slim==1.39.0
pydantic-graph==1.39.0
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.13.3
realtime==2.4.2
referencing==0.36.2
requests==2.32.3
RestrictedPython==8.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
s3transfer==0.11.4
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
scipy==1.15.2
SQLAlchemy==2.0.40
sse-starlette==2.2.1
starlette==0.46.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
tenacity==9.1.2
tokenizers==0.21.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
vecs==0.4.5
watchdog==6.0.0
wcwidth==0.2.13
websockets==14.2
wrapt==1.17.2
yarl==1.19.0
zipp==3.21.0
beautifulsoup4>=4.12
lxml>=5.0
